Deleting old thread_sys files
Threading is complicated, callbacks are annoying. I was trying to setup
a system where a manager thread would be in charge of spawning new
threads for tasks given to it and returning the outputs of completed
tasks, but this doesn't solve the callback hell that would ultimately
arise, and since it would be spawning any number of threads I just don't
feel safe with it. So, I will be moving over to async/await, which isn't
any easier BUT provides a decent solution to the callback hell problem.